requirements.txt
streamlit==1.35.0
streamlit-markmap==1.0.1
PyMySQL==1.1.1
openai==1.30.3
qianfan==0.3.13.1
python-dotenv==1.0.1